* The diskslice abstraction now stores offsets/sizes as 64 bit quantities.   (NOTE: DOS partition tables and standard disklabels can't handle 64 bit   sector numbers yet).  For future pluggable disklabel/partitioning schemes.
* The kernel panic / kernel core API is now 64 bits.

* The VN device now uses 64 bit sector numbers and can handle block devices
  up to what is supported by the filesystem (typically 8TB).  This change
  was made primarily so we can test future disklabel / partition table
  support.

* Pass 64 bit LBAs to various block devices and to the SCSI layer.

* Check for and assert 32 bit overflow conditions in various places, instead
  of wrapping.

+++ sys/bus/cam/scsi/scsi_all.h
@@ -15,7 +15,7 @@
  * Ported to run under 386BSD by Julian Elischer (julian@tfs.com) Sept 1992
  *
  * $FreeBSD: src/sys/cam/scsi/scsi_all.h,v 1.14.2.5 2003/08/24 03:26:37 ken Exp $
- * $DragonFly: src/sys/bus/cam/scsi/scsi_all.h,v 1.5 2006/05/21 03:43:38 dillon Exp $
+ * $DragonFly: src/sys/bus/cam/scsi/scsi_all.h,v 1.6 2007/05/15 22:44:02 dillon Exp $
  */
 
 /*
@@ -886,7 +886,7 @@ void		scsi_synchronize_cache(struct ccb_scsiio *csio,
 void scsi_read_write(struct ccb_scsiio *csio, u_int32_t retries,
 		     void (*cbfcnp)(struct cam_periph *, union ccb *),
 		     u_int8_t tag_action, int readop, u_int8_t byte2, 
-		     int minimum_cmd_size, u_int32_t lba,
+		     int minimum_cmd_size, u_int64_t lba,
 		     u_int32_t block_count, u_int8_t *data_ptr,
 		     u_int32_t dxfer_len, u_int8_t sense_len,
 		     u_int32_t timeout);
