+++ sys/dev/raid/vinum/vinum.c
@@ -37,7 +37,7 @@
  *
  * $Id: vinum.c,v 1.33 2001/01/09 06:19:15 grog Exp grog $
  * $FreeBSD: src/sys/dev/vinum/vinum.c,v 1.38.2.3 2003/01/07 12:14:16 joerg Exp $
- * $DragonFly: src/sys/dev/raid/vinum/vinum.c,v 1.19 2006/11/03 16:33:38 corecode Exp $
+ * $DragonFly: src/sys/dev/raid/vinum/vinum.c,v 1.20 2007/05/15 22:44:12 dillon Exp $
  */
 
 #define STATIC static					    /* nothing while we're testing XXX */
@@ -495,7 +495,7 @@ vinumsize(struct dev_psize_args *ap)
     vol = &VOL[Volno(dev)];
 
     if (vol->state == volume_up) {
-	ap->a_result = vol->size;
+	ap->a_result = (int64_t)vol->size;
 	return(0);
     } else {
 	return(ENXIO);
